requirements.txt
astroid==2.3.3
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.12
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
prettyconf==2.1.0
psycopg2==2.8.5
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
wrapt==1.11.2
